Move readDataFully check into FileWatcher.
This makes the metadata (last modified) that is sent to SyncLogic
more accurate (previously it was done post-SyncLogic), and also
matches how Facebook's Watchman works, which in theory would
make potentially moving to Watchman later easier.

It does mean we de-bounce ignored files, so it might be worth
eventually moving this into SyncLogic itself.