Fix bug where deleting multiple levels were echoed.
When a remote side would delete multiple levels, the initial
deletion would work, but when we heard the file system echos
(of our own deletes), this would re-populate the UpdateTree
with nodes.

These nodes were then considered newer than the remote, because
we'd purged all of the remote's updates when the delete happened.

The fix is to consider neither side "newer" if both sides are
deleted (or if we're deleted and the other side is null).