Fix directories getting echoed when files are added.
It turns out creating a new file within a directory sets
the directory's mod time to now.

Which is fine, but we don't need to consider this directory
now "newer" than the directory on the remote.

Alternatively, we could just not sync directories at all,
and only sync files+symlinks, and mkdir any directories as needed,
similar to what git does. But occassionally the "oh right, empty
directories aren't created" can trip up IDEs/etc., so it seems
easy/cheap to sync them for now.