Add a WatchmanFileWatcher.
Initially results look good.

The primary difference in semantics in that the WatchService API
would, for deletions, only send an event for the directory that
got deleted, and no events for it's contents.

Watchman sends results for every file affected by the change.

Which is fine, but it means we'll send useless updates over the
wire, e.g. if dir/ is deleted, we'll send deletes for both it
plus dir1/foo.txt, and other children.

Eventually the UpdateTree/UpdateTreeDiff should be able to
recognize when this happens, and not include those extra
deletes.