Minor updates for running in debug mode to watch FS event behavior.
Oddly, watchman sends events in a different order than the Java watcher.

Specifically, if I remove foo/ and it has foo/bar.txt in it, the order
I see in Java Watcher is:

* rm foo/bar.txt
* rm foo

Which makes sense, like the children are removed before the parent.

However watchman sends:

* rm foo
* rm foo/bar.txt

This should not be a big deal, but does mean a code path of "ensure
children were deleted" will happen for watchman all the time, because
the children events are just next in line.