Give a nice error message if the remote path does not exist.
This avoids a very unintuitive experience where the server is trying
to create paths that look valid, e.g. /Users/you/.../, but on the
server, e.g. Linux, the path should be /home/you/.../.

Also add a note in the readme and clarifications on relative vs.
absolute paths in the help docs.