Check for >500ms drift between the client and server.
If the clocks are too much out of sync, weirdness will result,
where one-side thinks it's side is newer when it's really not.

In theory should also add some safeguards to prevent mirror
from ever truncating files, when these sort of bugs/boundary
cases happen. (Which for me is basically never, but it's still
a bad thing when it does happen.)

Also should add a unit test for this.

Fixes #4.