Normalize the symlink before calling relative.
If the symlink is not normalized, e.g. it has foo/./bar, then it
will throw off the relative algorithm and return a base at the
first difference, e.g. ..././..., even though that is not actually
a different path.

Seems like this is a JVM regression because this was working before.