Split prevoius --debug into --debug-all/--debug-prefixes options.
The old option was confusing because running just `--debug` is a fairly
common idiom but, in the case of mirror, it wouldn't do anything because
it was seen as no prefixes being passed.

Now it should be more explicit which to use, and when to pass paths.

Also added a section to the readme about the help sections.