Migrate WatchmanFileWatcher to watch-project.
The watch command is deprecated, as watch-project plays nicer with
several separate tools using watchman to watch slightly different paths.

To do this, we have to differentiate between our root and the watch
root, and then: use the watch root for any query/subscribe commands
(with appropriate filters for only our subpath of the watch), also
strip our subpath off from the watch path when files come in.