Migrate to watchman-lib.
This was slightly annoying as watchman-lib is not published to
Maven central, and uses Buck to build, which doesn't build
on JDK 10/11 (iirc).

So, for now this uses a hand-published version of watchman-lib
that I've put into the repo.joist.ws repository.

The upshot is that we get to drop the hand-maintained cross-process
communication code, and also the watchman-lib already handles
de-multixplexing watchman responses back to N subscription listeners.

We don't technically use multiple subscriptions at once, but it
does mean we can kill/restart a subscription when overflows happen
without creating a brand new process, which is what we did previously.