Handle restoring deleted files with original timestamps.
Previously a restored file that had its original timestamp would not
be recognized as new, because the delete marker had its original
timestamp + 1 second.

This fix recognizes this scenario, and just bumps the original
timestamp to be after the delete marker.

This will change the modtime, but given mirror's current modtime-based
approach, that is necessary. In theory we could introduce a separate
timestamp field on paths that is ~kinda/sorta modtime but in situations
like this could be newer than modtime, but avoiding that for now.

Fixes #26.